Update FA1.2
Problem:
1. FA1.2 is not fully usable in current version of Michelson.
One can not call methods of an abstract FA1.2-compliant contract from
another contract because it requires knowledge of concrete parameter
type which is not specified in the document.
2. FA1.2 has a composability issue: if there appears a new standard
interface with a similar requirement on its parameter, it will become
impossible for a contract to comply with both requirements at the same
time.
3. There are some typos in FA1.2 description and some entrypoints are
described in subjectively bad way.

Solution:
1. Update "Approvable Ledger Interface" so that it only requires
presense of certain entrypoints, but does not require anything
regarding their position in the parameter tree.
2. Describe a temporary strict version of this interface which makes it
usable in Athens. We call it FA1.2.1
3. Ensure consistent spelling of "entrypoint(s)", that's how this word
is spelled in Michelson docs.
4. Rephrase descriptions of some entrypoints, fix some minor typos and
mistakes.