Remove no longer required FA1.2.1
Problem: Babylon upgrade introduced entrypoints feature that lifts
some of the restrictions and makes it possible to standardize
extensible interfaces. FA1.2.1 used proxy contract approach which
is expensive in terms of gas, and no longer required due to the
network upgrade. Leaving this standard in the repository might
cause confusion, and would require us to maintain outdated
recommendations.

Solution: Remove FA1.2.1 in favor of FA1.2, update the documents
accordingly.